Remove old data classes and associated workflows
